+++ packages/router-core/src/index.ts
@@ -115,6 +115,7 @@ export { encode, decode } from './qss'
 export { rootRouteId } from './root'
 export type { RootRouteId } from './root'
 
+export { BaseRoute, BaseRouteApi, BaseRootRoute } from './route'
 export type {
   AnyPathParams,
   SearchSchemaInput,
@@ -182,10 +183,10 @@ export type {
   RouteLoaderFn,
   LoaderFnContext,
   RouteContextFn,
-  RouteContextOptions,
   BeforeLoadFn,
-  BeforeLoadContextOptions,
   ContextOptions,
+  RouteContextOptions,
+  BeforeLoadContextOptions,
   RootRouteOptions,
   UpdatableRouteOptionsExtensions,
   RouteConstraints,
